cleanup seq API, now keyed-seq and simple-seq
as it will be very important in apps to choose the correct variant
the function call should be explicit instead of relying on
whether the second argument is nil or not.

now just need to write three pages of explanation about when to
use which. TL;DR: simple for things that don't change order and only
remove/add at end. keyed for pretty much all other cases.