rewrite dom event handling, test web component compatibility
event impl was bad now its a little bit better. not entirely sure
this is the best way to handle this but seems to align with what
I want (ie. no event functions in declarative parts)

added some tests for web component support. had to change
some attribute handling but that was overdue anyways. now 100%
compatible with regular webcomponents and DOM elements.