move idents to their own type instead of vectors
easier to keep track of and identify. avoids ambiguity with things
like [:foo :bar] looking like a ident to the db normalizer, when
it may just be a vector with 2 keyword elements.

also added default reader/writer impls for transit/edn so the
backend can easily decide how it wants to handle idents