add [:div (css ...) "foo"] shorthand for fragments
a large number of elements only contains {:class (css ...)} as
attributes, which gets visually noisy for no reason.

the fragment macro can infer that this is meant as the :class
attribute. assumes (css ...) is a shadow-css like macro actually
returning a classname.

this will break something (css ...) is actually something else
but that should be fine.

just turns into [:div {:class (css ...)} ...] and will merge
attributes in case (css ...) is followed by a regular attr map.